feat(api): moved patient parsing to core
refs. metriport/metriport-internal#1350
