+++ packages/core/src/external/carequality/pd/process-inbound-pd.ts
@@ -2,17 +2,21 @@ import {
   InboundPatientDiscoveryReq,
   InboundPatientDiscoveryResp,
 } from "@metriport/ihe-gateway-sdk";
-import { Patient } from "../../../domain/patient";
+import { InboundPatientResource } from "@metriport/ihe-gateway-sdk/src/models/patient-discovery/patient-discovery-responses";
+import { Address } from "../../../domain/address";
+import { Patient, PatientData } from "../../../domain/patient";
 import { MPI } from "../../../mpi/mpi";
 import { patientMPIToPartialPatient } from "../../../mpi/shared";
 import { toFHIR as convertPatientToFHIR } from "../../fhir/patient";
-import { validateFHIRAndExtractPatient } from "./validating-pd";
 import {
-  XDSRegistryError,
-  constructPDNoMatchResponse,
-  constructPDErrorResponse,
   IHEGatewayError,
+  XDSRegistryError,
+  constructPDErrorResponse,
+  constructPDNoMatchResponse,
 } from "../error";
+import { validateFHIRAndExtractPatient } from "./validating-pd";
+
+import { getStateEnum } from "../../../domain/geographic-locations";
 import { METRIPORT_HOME_COMMUNITY_ID } from "../shared";
 
 function constructMatchResponse(
@@ -54,4 +58,55 @@ export async function processInboundPatientDiscovery(
       );
     }
   }
+}
+
+export function patientDataFromResource(
+  patientResource: InboundPatientResource | undefined
+): PatientData | undefined {
+  if (!patientResource) return;
+  const humanName = patientResource?.name;
+  if (!humanName) return;
+  const firstName = humanName[0]?.given?.join(" ");
+  const lastName = humanName[0]?.family;
+  const dob = patientResource.birthDate;
+  const genderAtBirth = mapGender(patientResource.gender);
+  const addresses = getPatientAddresses(patientResource);
+
+  if (!firstName || !lastName || !dob || !genderAtBirth) return;
+  if (!addresses.length) return;
+
+  return {
+    firstName,
+    lastName,
+    dob,
+    genderAtBirth,
+    address: addresses,
+  };
+}
+
+function mapGender(string: string | undefined): "M" | "F" | undefined {
+  if (string === "male") return "M";
+  if (string === "female") return "F";
+  return;
+}
+
+function getPatientAddresses(patientResource: InboundPatientResource | undefined): Address[] {
+  if (!patientResource) return [];
+  const addresses: Address[] = [];
+  for (const address of patientResource.address) {
+    const state = address.state ? getStateEnum(address.state) : undefined;
+    const line = address.line ? address.line.join(", ") : undefined;
+    const city = address.city || undefined;
+    const zip = address.postalCode || undefined;
+    if (!state || !line || !city || !zip) continue;
+
+    addresses.push({
+      addressLine1: line,
+      city,
+      state,
+      zip,
+      country: address.country ?? "USA",
+    });
+  }
+  return addresses;
 }
